adding README gifs
